update user role after onboarding
