+++ src/components/onboarding/Onboarding.tsx
@@ -1,13 +1,9 @@
 'use client'
 
-import { NOT_AUTHORIZED } from '@/services/permissions/check'
-import { onboardOrganizationCommand } from '@/services/serverFunctions/organization'
-import { OnboardingCommand, OnboardingCommandValidation } from '@/services/serverFunctions/user.command'
-import { zodResolver } from '@hookform/resolvers/zod'
-import { Organization, Role } from '@prisma/client'
+import { Organization } from '@prisma/client'
+import { SessionProvider } from 'next-auth/react'
 import { useSearchParams } from 'next/navigation'
 import { useEffect, useState } from 'react'
-import { useForm } from 'react-hook-form'
 import OnboardingModal from './OnboardingModal'
 
 interface Props {
@@ -16,8 +12,6 @@ interface Props {
 
 const Onboarding = ({ organization }: Props) => {
   const [open, setOpen] = useState(false)
-  const [step, setStep] = useState(0)
-  const steps = 2
 
   const searchParams = useSearchParams()
   const onboarding = searchParams.get('onboarding')
@@ -27,50 +21,12 @@ const Onboarding = ({ organization }: Props) => {
     }
   }, [onboarding])
 
-  const form = useForm<OnboardingCommand>({
-    resolver: zodResolver(OnboardingCommandValidation),
-    mode: 'onSubmit',
-    reValidateMode: 'onBlur',
-    defaultValues: {
-      organizationId: organization.id,
-      companyName: organization.name || '',
-      role: Role.ADMIN,
-      collaborators: [{ email: '' }],
-    },
-  })
-
-  const previousStep = () => setStep(step > 1 ? step - 1 : 0)
-
-  const onValidate = async () => {
-    if (step < steps - 1) {
-      setStep(step + 1)
-    } else {
-      const values = form.getValues()
-      values.collaborators = (values.collaborators || []).filter(
-        (collaborator) => collaborator.email || collaborator.role,
-      )
-      const isValid = OnboardingCommandValidation.safeParse(values)
-      if (isValid.success) {
-        const res = await onboardOrganizationCommand(isValid.data)
-        if (!res || res === NOT_AUTHORIZED) {
-          onClose()
-        }
-      }
-    }
-  }
-
   const onClose = () => setOpen(false)
 
   return (
-    <OnboardingModal
-      open={open}
-      onValidate={onValidate}
-      form={form}
-      onClose={onClose}
-      activeStep={step}
-      steps={steps}
-      previousStep={previousStep}
-    />
+    <SessionProvider>
+      <OnboardingModal open={open} onClose={onClose} organization={organization} />
+    </SessionProvider>
   )
 }
